Pygame-RPG 1 The beginning stages
I started this project by importing the work done on an older text based
edition of the current project. The purpose of this project is to
upscale the olde work with Pygame while adding in new features.

These stages mostly consist of me using free assets I found online in
tandem with Pygame and seeing what the library could do. I mostly
fiddled around with the knight character's animations and seeing how the
background environment looked.