Pygame-RPG 3 Making dictionaries for the dungeons
This part is based on the dungeon maps I had drawn up for the older
version of this RPG. I decided that using dictionaries for this would be smarter
than using multiple if statements like I did before. This will make the dungeons
easier to test. The dictionaries i.e. dungeon maps that are used will be
set at the start  of every dungeon. This is so the dungeon map does not
change since the dictionaries will be how the player advances.

Although the events like finding an item or interacting with objects
aren't done yet, I placed strings detailing them so that when I can come
back I don't have to look at my hand drawn map. The MapInfo class in Map.py is
similar to the Managers I've made previously in that it acts differently
depending on the context. In this case, the context is which dungeon the
player is in.

If the player is in the Minotaurs dungeon it will contain information
about the minotaur dungeon's events. It also has whether the player has
accessed an event and whether there is an event on that space. If the
player accessed the event, the MapInfo object status will change from UNACCESSED
to ACCESSED which will make sure that the player can't view the same
events multiple times.