Pygame-RPG 5 Starting the tests for the managers
This commit will be the strt of me writting the tests for the manager
classes. These tests will be ran with each commit to ensure I don't
break anything. It will be the safety net I'll rely on so that I don't
have major problems in the future. For now since I do not fully
understand how python imports files from other directories, I've opted
to clone the manager files to the pytesting directory for now.

If I'm able to import the managers into rpg2.py while they are in the
pytesting directory then I will do just that. This commit also has the
start of my attempt to streamline variable names. For now I've chosen
that classes will be represented with each word starting with a
capitalized letter. There will be a future commit to fully address this
though.