Pygame-RPG 6 Moving Managers to manager directory
After many... many minutes of deep contemplation I come to the
realization that I miss Golangs package management. Nevertheless I
learned how python deals with packages and applied it. All manager
classes with the exception of Animation_Manager.py. The reason for this
is that anything relating to the player object AKA the Knight will be in
a separate package. As for Map.py, I plan to make a package called
UI to transmit information like that to the player so when that's done
it will be moved as well. Each package will have their own testing in
their directory moving forward.

Changes to Rpg2.py were made due to the movement of the manager classes.
Rpg2.py now imports the managers package and accesses the classes from
the package.