Pygame-RPG 6.1 Creation of PR template
Moving forward I will be making pull requests into master as another
layer of security. As such, it will be tiresome to have to remake
certain things. As such, this change will include the creation of the
.github directory and the pull_request_template file.

The pull_request_template file includes the following format:

Title
Summary
~
~
~
PR checklist
~
~