Pygame-RPG 7 Upgrading Manager Classes
Before this change, the manager classes used 'if' statements in order to
properly control what was occuring on the screen. This however, is not
efficient since when the scope of the project grows, so will the amount
of 'if statements. Instead of using 'if' statements, I elected to use
Python dicts.

The dicts that each manager uses are defined within their own files which
will make it easier to debug. I also took this time to fix some...
questionable programming decisions I had made to acomodate the old system.
The code in the rpg2.py file is now more streamlined.

As a result of using python dicts the previous ani_array variables now store
tuples instead of arrays. As such, they have been renamed ani_tuple and
their `change_array` function has become `change_tuple`. The
reason for this change is because lists cannot be returned by a dict but
tuples can. The reason is because lists can grow in size while tuples
are more like standard arrays.

Another change is that the context variable in some of the managers has
been removed and now only ScreenManager has a context variable. This
change is to make it easier to centralize everything into a single
omni-manager made from ScreenManager in the later stages of this project.

Some tests have needed to be tweaked due to the changes and I've
confirmed that the changes work without issue for now.