Pygame-RPG 8 Creating and Testing Dialogue Manager
parent dca05a2a869a1ffc3cea8ff65af868828467d7b3
author Keith Compere <keithcompere150@gmail.com> 1683560079 -0400
committer Keith Compere <keithcompere150@gmail.com> 1684802583 -0400

Dialogue Manager is the class that is responsible for managing the
dialogue in the game. The Dialogue Manager class has the logic for how
dialogue flows in the game. The `draw_dialogue()` function actually
draws the dialogue on screen for the player. Currently this is in the
form of a set timer, in the future, I would want this to be in the form
of the player pressing the 'Enter' key but I'm having trouble with that
implementation.

Portraits for the characters are found via a python dict and the input
for this dict is parsed from text found in a file. If a portrait is
found then `draw_dialogue()` will draw a character box and the portrait
inside the character box. The trigger for a text dialogue is through an
Event object through the eventType parameter. In the future, all player
interactions might be handled by a 'handle_interactions()' function for
the knight object which will combine all the logic for the interactions
into one place.

The textEnable variable, determines whether `draw_dialogue()` can be
used. textEnable is set to true when a dialogue Event object is
interacted with and is set to False when the dialogue event is over,
flipping the Event's activated variable to True in the process.

All the portraits and the pictures used for dialogue manager is stored
in the portraits directory. This change also includes the addition of a
brief description of the Dialogue Manager class.