Pygame-RPG 7.1 Upgrading Manager Classes
In the last PR I forgot to streamline the logic for the
Screen_Manager.py file. This PR addresses that oversight. In this PR,
the 'Event' class is also created in Screen_Manager.py which will
hopefully make the task of creating a dialogue manager easier for me in
the future. With the existence of the 'Event' class, the logic for
testing needed to be changed for the Object_Animation_Manager_test.py
file.