Pygame-RPG 9 SaveManager class creation
Every game needs a way to save. Although my game will be relatively
short, I still think everyone would appreciate a way to save the game.
Initially, I planned on using a local database(sqlite) but I realized
that parsing through a database for that information would be...
annoying. As such, I decided to store the information I needed in json
format.

The Save_Manager.py file houses the SaveManager class. SaveManager takes
in a Knight object and a dictionary of the local variables by using the
`vars()` function. It also has a limit of 5 saves and a saveNumber
variable which tracks the latest slot used. Upon creation, the object
will scan the save directory and look for a file called
`prev_save_info.json` which will mention which slot was used last. If
the `prev_save_info.json` file exists, the object will open the file and
read which slot was was last used and attempt to open that file.

The SaveManager class has 3 'main' functions. The first is
`strip_non_json_and_save()`. Since the dictionary of local variables
given to SaveManager is dynamic(it get's updated) it get's filled with
information that cannot be parsed to a json file using `json.dump()`. As
such, this function uses a list of allowed keys and creates a new dict
called 'rawVarsDict'. It also creates dictionaries for the Knight Object
and stores the context and the objects_dict which contains the events
for that screen(Name needs to be updated) from the ScreenManager given
to it into another dict. The dictionaries created are then returned together
in a dictionary.

The second functions of note is `quick_save()` and `save()`.
`quick_save()` and `save()` are identical in function with only one
difference that I will highlight later. The two functions both take in
a ScreenManager object and then create a file that will be written too
using `json.dump()`. The difference between `quick_save()` and `save()`
is that `quick_save()` saves the information in the last used save while
`save()` takes in a specified slot. Both versions update the
prev_save_info.json file

The final functions of note is `quick_load()` and `load()` functions.
The two read the information in a json save file and then update the
objects that require updating such as the knight and the ScreenManager
class. This includes updating the local variables stored in a dictionary in
SaveManager. The difference between `quick_load()`and `load()` is that
`quick_load()` reads from the latest save file while `load()` reads from
a specified slot.

Other changes:
- Knight classes received the `load_dict()` function to help load the
  information from the json file into the object. the knight class also
  received the `__eq__` function to compare knight objects (for testing)

- Dummy_Knight class updated to behave like the actual Knight class for
  testing purposes. This includes giving it a way to level up for it's
  stats to change.

- save directory and 3 mock json files added to repo for jenkins to test
  with

- tests for the SaveManager class have been written

- `ScreenManager.change_screen()` no longer returns anything and the
  tempScreen variable has been replaced with calls to
  `ScreenManager.screen`

- Temporarily added a shortcut to saving and loading by pressing s and
  then l on the keyboard(This will be removed later). Also added a line
  to update NPCAnimationManager when loading from a save file.

- Created a script file that will house useful bash scripts(currently is
  only called to clear the save directory after tests)

- Refactored the Knight.status value to use a string instead of an Enum
  since Enums weren't Json compatible