Pygame-RPG 9.1 Housekeeping and bugfixing
Due to an oversight, the code responsible for stopping the player from
going out of bound was malfuntioning. What happened was that due to
some other changes made in the Pygame-RPG 9 PR, the
`game.image.load()` function was being called whenever the `change_screen()`
function was invoked. Since function would put the player right on the
border of the screens, that meant that it could be invoked multiple
times per second if the player kept going in that direction.

This lead to a drop in framerate that resulted in a slowdown in the
animations playing on screen. As a result, some changes needed to be
made to files for this slowdown to stop happening. First of all,
functions in the ScreenManager class needed to be changed. The
`change_screen()` now has an if-else block to stop the
`game.image.load()` fuction being invoked everytime `change_screen()` is
invoked. Due to this however, I could no longer call `change_screen()`
in `change_context()` for ScreenManager objects. This is problematic
because that piece of code was responsible for rendering the new screen
when the player loads a save file. Instead of calling `change_screen()`
`change_context()` manually loads the new screen and also calling
`apply_context()` to load all of the other information for that screen.
As a result, `change_context()` no longer takes pos as an argument and
I've removed it from any function that calls it.

To prevent a disaster like this from happeneing, I've implemented a
framerate counter at the top left of the screen. This will help me
better find out when something similar goes horribly wrong.

Other changes are as follows:
- Changing the format of comments to match pycharm convention
- Simplifying equality expressions
- Other minor changes for convention to be followed
- Addition of some comments for future reference