Pygame-RPG 9.2 Housekeeping and bugfixing
It was discovered that the states of the events were saved by
SaveManager. This led to a glitch where you could open a chest, load a
previous save from before you opened the chest and be unable to open
said chest. As such, SaveManager needed to be changed to include this
vital feature. Tests for this were also written.

Other changes:
- created the `load()` function to simplify loading JSON data into the
  object
- created the `load_data()` function to simplify loading information in
  SaveManager
- formatting changes