Pygame-RPG 10 Creation of UIManager
UIManager is a class created for the sake of both drawing and handling
the UI elements of the game. This class is mostly going to be used for
handling menu's or choosing dialogue choices in the future.

The following describes how the `draw_UI()` function works for
UIManager. First, the UIManager looks to see if other assets(images) are needed
to draw the UI by putting the UI in a certain dictionary. If the
function given is not None, the function given is called. The functions
in the dictionary are in a file called UI_Manager_draw.py

UIManager then looks at the assigned restraints on both the X and Y axises as well
as the given X and Y spacing for that UI. The combination of those two
factors determines how the assets are drawn.

assets = ["Hi", "Hello", "OK"]
ex) if X spacing is 0

	Hi
	Hello
	OK

ex) if X spacing isn't 0 and X restraint allows for 2 items in X before it
is surpassed

	Hi Hello
	OK

Finally, the object makes a call to `handle_cursor()` and invokes the
cursor objects `handle_cursor()`. both of these `handle_cursor()`
functions take the keys array generated by `game.key.getpressed()`.

The cursor object then checks the input given and moves the cursor image
based on the restrictions that were passed to it by the
`set_cursor_info()` in the Cursor object.

`draw_UI()` can return 2 results. The first flow is the following.
If the player didn't press enter, the `handle_cursor()` function in the
Cursor object returns false. This means there's no need to process the
input further therefore the `draw_UI()` function returns None. If they did press
the enter key, the `handle_cursor()` in the Cursor object returns true.
an array of size 2 is returned. The first item is the UI in question and the
second item is the users choice which is found by comparing the current
position of the cursor, the X and Y restraints as well as the X and Y
spacings to find the option chosen. this array is then returned
by `draw_UI()`to be processed.

Other changes:
- Creation of the Scripts.py file
- Some slight changes to SaveManager to make it more reflective with the
  UI
- removed save_data0.json and replaced it with save_data1.json to pass pytests
- Added some PNG assets that are used by UIManager