Pygame-RPG 10.1 Housekeeping and bugfixing
It has been a long time since I've updated my README.md so I decided to
work on that for the bulk of this commit. Seeing that my README.md was
getting too content heavy, I divided them into two separate READMEs. The
one at the root of the project is the README with basic information
about the game while the README in the managers folder has the more
technical details about what each class does.

I also decided to finally delete the test.py file from GitHub to stop it
from being tracked. Since I just use the file for testing before I move
to integration there was no real need for the changes on it to be
tracked anyway.