Pygame-RPG 11 Creation of UIHandler
This Commit focuses on creating the UIHandler class which handles the
interactions the player has with the UI and translates those choices to
actions taken in the game. As a result, the UIHandler class takes holds
a reference to various Manager objects in order to do this.

Objects UIHandler has a reference to:
- UIManager: For operations involving moving from one UI to the next
  based on player choice.

- SaveManager: For Save and Load operations done by the player.

- Knight: Not in use for now but for the sake of item use and battles
  this reference may be used.

- localVars: Some UI Interactions require that local variables be
  changed as a result. Example, setting the `start` variable to false
  once the player chooses an action that would make them leave the start
  screen.

- BattleManager: A not yet implemented class but since BattleManager
  will have a UI component this is neccesary.

The UIHandler class works as follows. The classes' main function is the
`handle_interaction()` function which takes in a context and choice
gotten from UIManager and processes the action required for that
specific context and choice. These actions range from changing the UI to
directly manipulating local variable values. As always, a file testing
the utility of the UIHandler class was created.

This commit also aims to integrate the UIManager and UIHandler classes
into the Rpg2.py file so that they can be play tested. As a result,
there is a starting loop that depends on the `start` boolean variable
that sets the UI to the start screen and forces the player to stay there
until they pick a choice that switches `start` to false. If the 'start'
variable is switched to false because of the choice 'Start Game' then
the screenManager context is manually changed, or else it will be
handled by the load operation the player chose.

Some behavioral changes were also made to certain files, the changes
were the following.

- UIManager's title for the 'Load Game' UI is changed to 'Load Game' and
  it's position was adjusted accordingly

- If None is passed to `UIManager.change_UI()` the object will be
  'cleared'

- `Save_Manager_test.py` had the range of it's random value for
  animationTracker3 reduced since it was possible to get an invalid
  number for this stage of development

- As a result of the 'invalid number' problem mentioned earlier, the
  `save_data.json` files were updated with non-problematic files. I do
  find it weird how this bug just started happening now though... It
  could have just been blind luck though.