Pygame-RPG 10.2 Housekeeping and bugfixing
This series of housekeeping changes is aimed at simplifying the overall
structure of the project. These simplifications mostly include the
addition of functions to the Rpg2.py file. The following functions were
created for these purposes:

`handle_basic_input()`: takes in keys, knightAni, knight object, x and
animationTracker to handle the basic movement of the knight character.
returns x position of knight and animationTracker

`handle_player_interaction()`: takes in keys, knight object, various
managers along with textEnable and animationTracker3 to handle the
knight objects interactions with events. For now this is only in the
form of opening a chest. Returns textEnable and animationTracker3

`draw_objects()`: takes in screen, screenManager and spot3. This
function is responsible for drawing any and all objects that appear on
the screen.

`change_screen()`: takes in x position, screenManger and NPCManager.
This code is responsible for the switching of screens.

These changes have made dealing with the Rpg2.py file simpler and easier
to handle. Another change to the Rpg2.py file is the fact that
the `draw_dialogue()` function is now a part of Dialogue_Manager.py.

Other changes:
- DialogueManager class now uses events from Pygame event queue to work
  instead of a timer. This also allows only one 'enter' press at a time.
  This leads to a smoother experience overall.
- Changes made to Dialogue_Manager.py are reflected in
  Dialogue_Manager_test.py
- SaveManager class now has a reference to any manager it saves data from
  so it does not have to pass in managers for `save()` and `load()`
  functions.
- Save_Manager_test.py now reflects the changes made in Save_Manager.py
- Scripts.py's fill-save functionality now reflects the changes made to	          Save_Manager.py
- Added the proper flag for fill-save option in script file. When
  `script fill-save` is called, it now calls `Scripts.py fill-save` instead
  of just `Scripts.py`