Pygame-RPG 10.3 Housekeeping and bugfixing
This commit enhances the cursor movement by using the Pygame event Queue
instead of using the `game.key.get_pressed()` command. This lets the
system 'chain' inputs in a smooth manner thus eliminating the
requirement for a time delay in the code. The code that handles
processing the events does break when the player enters 'enter' or
'esc' because those keys can trigger a change in UI.

The tests in Dialogue_Manager_test.py were also changed to reflect the
new event queue system. All instances of the `ensure_pressed()` command
has been replaced with a new command called `get_keydown_event()`. This
command takes in a string key and returns an event queue that involves
a created a Keydown event based on the inputed key.
example, inputing the key "w" gives an event queue with a Keydown event
with a key value of `game.K_UP` etc.

As a result of these changes, the script files run-tests flag has been
modified to only run the `pytest managers` command. This is because, in
the advent of these changes, it is no longer required to create a file
containing the test data. Incidently, the `run-tests` flag has been
removed from the Scripts.py file since it is now unnecessary.