Pygame-RPG 13.1 Preparation for BattleManager class
This PR is the first of the preparatory PRs before work on the
BattleManager class can be worked on. This PR focuses on implementing
methods and classes to streamline Entity relationships. As a result, 2
new classes have been made, `Entity` and `Enemy`.

Classes Added:
- Entity: The entity class serves as the base model for the `Knight` and
  `Enemy` classes to take from. It's `__init__()` function contains the
  stats that both classes share. It also contains base methods that both
  classes need.
  These base functions are as follows:
  - `attack()`: which hasn't been written yet but is a basic attack
    function.
  - `take_damage()`: reduces incoming damage by the defence stat and
    then applies it to the object. This function checks if the attack is
    fatal(AKA Hp is <= 0 after the attack).
  - `apply_bonuses()`: Applies stat buffs that the entity has,
    decrements the buff counter if the buff isn't an infinite duration
    buff.
  - `remove_bonuses()`: removes applied buffs at the end of turn to
    prohibit lingering buff effects. Also removes buffs that have run
    out(the buff count == 0).

- `Enemy`: The enemy class inherits from the Entity class and
  is the template for all the enemy type
  objects that will appear. The templates for the objects that this
  class represents will be represented in JSON and they will be created
  by a factory type class. This class also has the `uniqueBuff` parameter
  which is meant to be used by bosses. The `weakness` parameter is a list
  of attributes that enemy is weak too and will thus take extra damage
  from
  the functions added to this class are the following:
  `die()`: This function adds the dropped items, experience value and
  the money the enemy drops into the loot pool. This loot pool is really
  just a dictionary that tracks the 'loot' from defeated enemies to be
  given to the player upon victory.

The Knight class now inherits from the Entity class. The Knight class
now has the `level_up()` function and the `get_rewards()`. The
`level_up()` function hasn't been written yet but it will take a list of
growths and apply those growths to their respective stats. It then
subtracts the Exp value by the Expcap value. It then increases the
threshold for the level up. The `get_rewards()` function simply gives
over the rewards from the loot pool in a battle. It calls `level_up()` if
Exp exceeds or is equal to the knight's Expcap.

As a result of all of these changes, I've created the Entity package for
classes related to Entity management.