Pygame-RPG 13.2 Preparation for BattleManager class
This PR is the second of the preparatory PRs before work on the
BattleManager class can start. This PR focuses on implementing the use
of JSON information into classes. As a result, JSON files have now been
added for most manager classes and the `Entity_Factory.py` file has been
created.

Classes Added:
- EntityFactory: The EntityFactory class is responsible for creating
  entities for future use based on an established JSON template. It
  takes in no parameters and only has a single attribute, `path` which
  is a static path to the JSON files it needs. This class only has one
  method for now.
  - `create_entity()`: This function takes in the name of the entity in
    question and appends it to the path variable and adding the ".json"
    ending to point to a specific file. If the file exists, it's
    information is loaded into a dictionary. This dictionary is then
    passed into the constructor for the `Enemy` class, returning an
    `Enemy` object.

As a result of how the EntityFactory class works, changes were also made
to the `Enemy` class. Instead of taking in each individual parameter,
the `Enemy` class constructor is now designed to take in a dictionary.
It passes the information over to the `Entity` constructor as
well as set it's own unique attributes based on the dictionary's
information.

Another change that was made was the creation of a new script for my
personal use. Now the script file has a `create-enemy` function that
creates a simple enemy template of a specific name. This makes it so I
can just automate the creation of JSON templates for the `Enemy` class.
The actually executed code is in the Scripts.py file.

The final major change is that most manager classes now take dictionary
variable information from JSON files, making it easier to change things and
reducing the clutter in the `Manager` class files. Some dictionaries
however could not be turned to JSON data for various reasons.
As a result of these changes, most manager class files first start by
taking the dictionary information from JSON files and putting it into
their variable space. Further work on this topic will be done in
housekeeping PRs from now on.

Other changes include some slight naming corrections and other minor
changes to the files such as `Rpg2.py`. Just like the previous PR, all
testing files will be created in the final PR.