Pygame-RPG 13.4 Housekeeping and bugfixing
This commit updates the README.md files in managers package and at the
root of the project and created the README.md file in the Entity
package. This README.md file explains the basic use of the classes in the
Entity package along with a quick introduction to how their functions work.
There was also a slight change to the `Entity/Knight.py` and
`managers/Dummy_Knight.py` file's `load_dict()` function. It was simplified to
simply take in the new dictionary as it's own rather than parse individual
values.