Pygame-RPG 15 Battle UI elements implementation
This commit focuses on all of the changes made to the
UIManager and UIHandler classes that allow for the UI in the battle to
function. The BattleManager class also had some changes that tweak it's
behavior. As a result of this, assets were added in and the two classes
were modified to accommodate the new system.

The changes made to the UIManager class are the following.
The attributes, `targets`, `targetSlider`, `subMenu`, `subMenuItems`,
`subMenuMinIndex`, `subMenuMaxIndex`, `subMenuSlider` were added. The
`draw_submenu()`, `handle_submenu_input()` and `handle_select_inputs()`
functions were created. As a result of these functions, the `draw_UI()`
and `handle_cursor()` functions were tweaked to fit the desired
behavior.

Attributes added:
`targets`: A list of tuples that contains the x and y positions of the
targetable entities. This can never be an empty list
`targetSlider`: An integer that represents which location the cursor is
currently pointing to.
`subMenu`: A boolean value that indicates whether a submenu is active or
not.
`subMenuItems`: a list of string values that are to be displayed on the
screen.
`subMenuMinIndex`: Since a submenu can only display 9 items from
`subMenuItems`, this integer tells us the lowest index for the
displayed items.
`subMenuMaxIndex`: Since a submenu can only display 9 items from
`subMenuItems`, this integer tells us the maximum index for the
displayed items.
`subMenuSlider`: This integer tells us which item the player currently
has the cursor on.

Functions Added:
`draw_submenu()`: This function draws the graphical portion of the
sub menu and populates this with the information from
`subMenuItems`using `subMenuMinIndex` and `subMenuMaxIndex` as the
boundary for the displayed items.
`handle_submenu_input()`: This function takes in a pygame event list and
handles arrow key, escape key and enter key events. It then processes
this for submenus since there is a different logic being applied for the
movement of the cursor in submenus. Instead of moving the cursor based on
pre-established boundaries, this code moves the `subMenuSlider` value
through the array differently depending on the arrow key pressed.
`handle_select_inputs`: This function takes in a pygame event list and
handles arrow key, escape key, and enter key events. It then processes
this for moving the `targetSlider` integer through the `targets` list.

The `handle_cursor()` functions were changed due to the addition of
submenus and targetting. This function takes in a pygame event list and
instead of processing it using the Cursor class's `handle_cursor()`
function, it is now split into 3 separate scenarios. If a submenu is
active and the context of the UIManager isn't "Select Target",
`handle_submenu_input()` is invoked and the code runs checks on
that scenario, this includes ensuring that the `subMenuSlider` value
isn't in an illegal state (an integer that doesn't make sense
considering the list and min and max indices). If the context is "Select
Target" then `handle_select_input` is invoked and the code ran ensures
that `targetSlider` isn't in an illegal state. Finally, if it fufills
none of the previous conditions, the Cursor object's `handle_cursor()`
function is invoked.

Depending on which pathway, the code goes through,
the cursor is drawn differently. In the case of a submenu, the cursor
drawing location is determined via some math using `subMenuSlider`,
`subMenuMinIndex` and `subMenuMaxIndex`. When the context is "Select
Target", the cursor is drawn using the tuple from `targets` and the
other case is the normal case.

In the case of UIHandler, the following were changed. Some new values
were added to the file for testing purposes and the class can now handle
subMenu and targeting cases aside from the other cases. It now also
properly takes in a battleManager object for the sake of the targeting
context.

Other Changes:
- Knight gained a function that turns it's inventory dictionary into a
  string dictionary
- BattleManager gained the `add_enemy()` function that takes in a Entity
  object and adds it to the `enemies` list of tuples with an automatically
  generated x and y position tuple. ex ((500, 300), Entity)
- JSON files were updated to allow "Select Target" and "Player Select"
  context's to work
- UI_Manager_draw.py gained a new function that just draws the temporary
  background for player select
- UI Assets were added

Concerns:
This commit WILL break a lot of tests due to the aformentioned changes.
Normally I'd fix the test cases now but I unfortunately do not have the
time right now. The behavior for the changes however, has been manually tested and it lines up with the desired behavior